feat: improved test_neighbour function, removed iris_rf_classifier fixture and refactored tests
